Clean up use case implementation
The previous iteration of use cases could cleanly account for non-null
or non-existent parameters. We would be would have to force unwrap
the parameters to be able to use them inside of the use case without
compiler warnings.

This implementation also allows for code like this:
```
getUser.execute(userId).subscribe()
```
to be replaced with code like this:
```
getUser(userId).subscribe()
```